requirements.txt
ensure==1.0.2
ipykernel

yt-dlp==2023.3.4
pandas==1.5.3
numpy==1.23.5
openai-whisper==20230314
pathlib==1.0.1

-e .